Setting .children requires an iterable.
To have a fully transparent .children attribute, getter and setter need
to work on the same data structure. The data structure is an iterable
for setting and a read-only iterable (tuple) on reading.
So setting just one element as children would confuse the users as
they would expect a single children in that case on read.