Update dotexporter to work on Windows/Cygwin
On Windows, the current "to_picture" function does not work as the temporary file name is not visible to the system. Therefore, calling "dot" on the command line will result in a FileNotFoundError.

The fix changes the function so that on windows/cygwin, the file is not deleted until after the picture is made. Once the picture is made, the file in unlinked.