Use __slots__ on NodeMixin to reduce memory usage
The usage of `__slots__` is highly recommended in order to reduce memory usage on classes, even more for the ones which are used a lot of time as it could be `NodeMixin`. The benefit here is that `__dict__` is not generated then and a lot of memory is saved with this. Besides, with this change any class which inherit `NodeMixin` can now benefit of not creating `__dict__` and reduce the memory usage, otherwise child classes cannot reduce memory with this trick if the parent class does not implement it too.

However, the rest of the classes of this package which inherit from `NodeMixin` should not use `__slots__` because they benefit from the flexibility of `__dict__`. Hence, this benefit is only for implementations which use directly `NodeMixin`, as mine.

In order to set some statistics, in my implementation, the use of this change help us reduce the memory usage of the custom node classes by 49,91% and by 71,85%. Hence, as these classes are highly used on my implementation this can represent a lot of MBs reduction on memory usage.